# packages/shared/src/openApi/impact-cocoa-client/schema.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CocoaImpact API",
    "version": "1.0.0",
    "description": "ImpactCocoa Data Management Platform API"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Backend"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": ["error"]
      },
...
      "Farmer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cooperativeId": {
            "type": "string",
            "format": "uuid"
          },
          "cooperativeCode": {
            "type": "string"
          },
          "cooperativeName": {
            "type": "string"
          },
          "districtName": {
            "type": "string",
            "nullable": true
          },
          "farmerCode": {
            "type": "string"
          },
          "externalSource": {
            "type": "string",
            "nullable": true
          },
          "producerId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "otherNames": {
            "type": "string",
            "nullable": true
          },
          "sex": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "nationalIdNumber": {
            "type": "string",
            "nullable": true
          },
          "nationalIdType": {
            "type": "string",
            "nullable": true
          },
          "village": {
            "type": "string",
            "nullable": true
          },
          "section": {
            "type": "string",
            "nullable": true
          },
          "dataCollectionConsent": {
            "type": "boolean",
            "nullable": true
          },
          "certificationStatus": {
            "type": "string"
          },
          "registrationDate": {
            "type": "string",
            "nullable": true
          },
          "householdSize": {
            "type": "integer",
            "nullable": true
          },
          "childrenCount": {
            "type": "integer",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "cooperativeId",
          "cooperativeCode",
          "cooperativeName",
          "districtName",
          "farmerCode",
          "externalSource",
          "producerId",
          "firstName",
          "lastName",
          "otherNames",
          "sex",
          "dateOfBirth",
          "phoneNumber",
          "nationalIdNumber",
          "nationalIdType",
          "village",
          "section",
          "dataCollectionConsent",
          "certificationStatus",
          "registrationDate",
          "householdSize",
          "childrenCount",
          "isActive",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "FarmerDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "cooperativeId": {
            "type": "string",
            "format": "uuid"
          },
          "cooperativeCode": {
            "type": "string"
          },
          "cooperativeName": {
            "type": "string"
          },
          "districtName": {
            "type": "string",
            "nullable": true
          },
          "farmerCode": {
            "type": "string"
          },
          "externalSource": {
            "type": "string",
            "nullable": true
          },
          "producerId": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "otherNames": {
            "type": "string",
            "nullable": true
          },
          "sex": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "nationalIdNumber": {
            "type": "string",
            "nullable": true
          },
          "nationalIdType": {
            "type": "string",
            "nullable": true
          },
          "village": {
            "type": "string",
            "nullable": true
          },
          "section": {
            "type": "string",
            "nullable": true
          },
          "dataCollectionConsent": {
            "type": "boolean",
            "nullable": true
          },
          "certificationStatus": {
            "type": "string"
          },
          "registrationDate": {
            "type": "string",
            "nullable": true
          },
          "householdSize": {
            "type": "integer",
            "nullable": true
          },
          "childrenCount": {
            "type": "integer",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "deletedAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "cooperativeId",
          "cooperativeCode",
          "cooperativeName",
          "districtName",
          "farmerCode",
          "externalSource",
          "producerId",
          "firstName",
          "lastName",
          "otherNames",
          "sex",
          "dateOfBirth",
          "phoneNumber",
          "nationalIdNumber",
          "nationalIdType",
          "village",
          "section",
          "dataCollectionConsent",
          "certificationStatus",
          "registrationDate",
          "householdSize",
          "childrenCount",
          "isActive",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "CreateFarmerBody": {
        "type": "object",
        "properties": {
          "otherNames": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          },
          "sex": {
            "type": "string",
            "enum": ["male", "female", "other", "unknown"]
          },
          "dateOfBirth": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true,
            "maxLength": 32,
            "pattern": "^[+\\d\\s().-]+$"
          },
          "nationalIdNumber": {
            "type": "string",
            "nullable": true,
            "maxLength": 64,
            "pattern": "^[A-Za-z0-9-]+$"
          },
          "nationalIdType": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          },
          "village": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          },
          "section": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          },
          "dataCollectionConsent": {
            "type": "boolean",
            "nullable": true
          },
          "certificationStatus": {
            "type": "string",
            "maxLength": 200
          },
          "registrationDate": {
            "type": "string",
            "nullable": true,
            "format": "date"
          },
          "householdSize": {
            "type": "integer",
            "nullable": true,
            "minimum": 0,
            "maximum": 100
          },
          "childrenCount": {
            "type": "integer",
            "nullable": true,
            "minimum": 0,
            "maximum": 50
          },
          "isActive": {
            "type": "boolean"
          },
          "producerId": {
            "type": "string",
            "nullable": true,
            "maxLength": 200
          },
          "cooperativeId": {
            "type": "string",
            "format": "uuid"
          },
          "farmerCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "pattern": "^[A-Za-z0-9_-]+$"
          },
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[\\p{L}\\p{M}\\p{N}'\\-. ]+$/u"
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "pattern": "^[\\p{L}\\p{M}\\p{N}'\\-. ]+$/u"
          }
        },
        "required": ["cooperativeId", "farmerCode", "firstName", "lastName"]
      },
...
    },
...
  },
  "paths": {
...
    "/api/farmers": {
      "get": {
        "tags": ["Farmers"],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "q",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cooperativeCode",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "section",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "village",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "certificationStatus",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            },
            "required": false,
            "name": "isActive",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            },
            "required": false,
            "name": "includeDeleted",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated farmers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Farmer"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  },
                  "required": ["items", "total", "page", "pageSize"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Farmers"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFarmerBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/farmers/{id}": {
      "get": {
        "tags": ["Farmers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Farmer detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Farmers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFarmerBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Farmers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Soft-deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/farmers/{id}/restore": {
      "post": {
        "tags": ["Farmers"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Restored",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmerDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Not deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}